move report utility functions from types to utils module and update imports
